# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Baseurl /showme/v1
        
  
# during dev, should point to your local machine
host: localhost:80
...
basePath: /showme/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /leaders/{user_id}:
    x-swagger-router-controller: leader
    get:
      description: API for get leaders information by id.
      # used as the method name of the controller
      operationId: getLeaderInfo
      parameters:
        - name: user_id
          in: path
          description: The user's id.
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/BaseSuccessRespone"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"    
  /leaders/location:
    # binds a127 app logic to a route
    x-swagger-router-controller: leader
    get:
      description: API for get leaders location in range.
      # used as the method name of the controller
      operationId: leaderLocation
      parameters:
        - name: radius
          in: query
          description: radius which user input.
          required: false
          type: number
        - name: lng
          in: query
          description: longitude which user input.
          required: false
          type: number
        - name: lat
          in: query
          description: lattitude which user input.
          required: false
          type: number
        - name: address
          in: query
          description: address which user input.
          required: false
          type: string  
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/BaseSuccessRespone"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  #   put:
  #     description: API for update leaders location.
  #     # used as the method name of the controller
  #     operationId: updateLocation
  #     parameters:
  #       - name: leader
  #         in: body
  #         description: leader's info.
  #         required: true
  #         schema:
  #           $ref: "#/definitions/updateLocationParams"   
  #     responses:
  #       "200":
  #         description: Success
  #         schema:
  #           # a pointer to a definition
  #           $ref: "#/definitions/BaseSuccessRespone"
  #       # responses may fall through to errors
  #       default:
  #         description: Error
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  # /leaders/active:
  #   x-swagger-router-controller: leader  
  #   put:
  #     description: API for update leaders status.
  #     # used as the method name of the controller
  #     operationId: updateStatus
  #     parameters:
  #       - name: leader
  #         in: body
  #         description: leader's info.
  #         required: true
  #         schema:
  #           $ref: "#/definitions/updateStatusParams"   
  #     responses:
  #       "200":
  #         description: Success
  #         schema:
  #           # a pointer to a definition
  #           $ref: "#/definitions/BaseSuccessRespone"
  #       # responses may fall through to errors
  #       default:
  #         description: Error
  #         schema:
  #           $ref: "#/definitions/ErrorResponse"
  /comment:
    # binds a127 app logic to a route
    x-swagger-router-controller: leader
    get:
      description: API for get comment.
      # used as the method name of the controller
      operationId: comment
      parameters:
        - name: user_id
          in: query
          description: user_id of user.
          required: true
          type: string
        - name: role
          in: query
          description: role of userr.
          required: true
          type: string
        - name: type
          in: query
          description: type of query. 'all' for user profile/'one' for individual user.
          required: true
          type: string 
        - name: page_size
          in: query
          description: record per page.
          required: true
          type: number
        - name: page_number
          in: query
          description: page want get record.
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/BaseSuccessRespone"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /leaders/statistical:
    # binds a127 app logic to a route
    x-swagger-router-controller: leader
    get:
      description: API for get leaders statistical in range.
      # used as the method name of the controller
      operationId: leaderStatistical
      parameters:
        - name: user_id
          in: query
          description: user_id of leader.
          required: true
          type: string
        - name: page_size
          in: query
          description: record per page.
          required: true
          type: number
        - name: page_number
          in: query
          description: page want get record.
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/BaseSuccessRespone"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
...
definitions:
  BaseSuccessRespone:
    required:
      - message
      - returnCode
    properties:
      message:
        type: string
      returnCode:
        type: number
      data:
        type: object
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
      returnCode:
        type: number
      data:
        type: object
  #User params for /user post method      
...